refactor: 출석 패키지 분리
