+++ src/main/java/com/waruru/areyouhere/active/AuthCodeController.java
@@ -1,4 +1,4 @@
-package com.waruru.areyouhere.attendance.controller;
+package com.waruru.areyouhere.active;
 
 
 import com.waruru.areyouhere.attendance.service.rdb.AttendanceRDBService;
@@ -7,7 +7,7 @@ import com.waruru.areyouhere.course.service.CourseService;
 import com.waruru.areyouhere.session.domain.entity.Session;
 import com.waruru.areyouhere.attendance.dto.request.AuthCodeDeactivationRequestDto;
 import com.waruru.areyouhere.attendance.dto.request.AuthCodeRequestDto;
-import com.waruru.areyouhere.active.ActiveSessionService;
+import com.waruru.areyouhere.active.service.ActiveSessionService;
 import com.waruru.areyouhere.session.service.command.SessionCommandService;
 import com.waruru.areyouhere.session.service.query.SessionQueryService;
 import java.time.LocalDateTime;
